requirements.txt
flask
google-generativeai==0.3.2
tensorflow==2.13.0
numpy
pillow==10.0.1
